changes services to snake case and refactors
